// src/app/models/voo.model.ts
export class Voo{
    vooId?:number;
    codigoAviao?: String;
    partida?: any;
    chegada?: any;
    totalAssentos?: number;
    status?: string;
    origem?:String;
    destino?:String;
    valorAssento?:number;
    economica?: string;
    executiva?: string;
    primeira?: string ;
    qtdAssentoPrimeira?: number;
    qtdAssentoExecutiva?: number;
    qtdAssentoEconomica?: number;
    precoAssentoPrimeira?: number;
    precoAssentoExecutiva?: number;
    precoAssentoEconomica?: number;
    horaChegada?:any;
    horaPartida?:any;
}